Refactoring.
Code cleanup.
Moved fuctions to utils folder (package).
Not tested but should work.
